refactor(types): types.rs 拆分为 types/ 目录模块
- 2828 行 → mod.rs 34 行 + git/usage/workspace/email/settings/providers/tests 七个子文件
- 纯搬运+桶重导出，93 个引用文件零改动；禁止改名纪律遵守
- email 两个 default fn 升 pub(crate) 供 settings.rs serde 引用（计划允许的显式路径方案）
- daemon 的 #[path] 指向 ../types/mod.rs
- baseline 重生成 + 新文件登记豁免
验证：cargo check 通过；cargo test --lib types 31 passed、--bin cc_gui_daemon types 24 passed；gate 绿
